+++ release/components/footer/footer.directive.js
@@ -2,7 +2,7 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 var core_1 = require("@angular/core");
 var footer_template_directive_1 = require("./footer-template.directive");
-var DatatableFooterDirective = /** @class */ (function () {
+var DatatableFooterDirective = (function () {
     function DatatableFooterDirective() {
     }
     DatatableFooterDirective.decorators = [