+++ release/datatable.module.js
@@ -6,7 +6,7 @@ require("rxjs/add/observable/fromEvent");
 var components_1 = require("./components");
 var directives_1 = require("./directives");
 var services_1 = require("./services");
-var NgxDatatableModule = /** @class */ (function () {
+var NgxDatatableModule = (function () {
     function NgxDatatableModule() {
     }
     NgxDatatableModule.decorators = [